requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cycler==0.10.0
efficientnet-pytorch==0.7.1
et-xmlfile==1.1.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
opencv-python==4.5.3.56
openpyxl==3.0.7
pandas==1.3.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
